sync upto PR#43
